Update final HMM in I-VDT-HMM to use duration/dispersion based classifier

+++ ternary_classifier/classificator_1.5/PDF_function.m
@@ -6,6 +6,6 @@
 function [ probability ] = PDF_function(x,mean, std )
 %PDF_FUNCTION Summary of this function goes here
 %   Detailed explanation goes here
-    probability = 1/sqrt(2*pi*std^2)*exp(-(x-mean)^2/(2*std^2));
+    probability = 1/sqrt(2*pi*std^2)*double(exp(-(x-mean)^2/(2*std^2)));
 end
